+++ packages/eslint-basic/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @yomukizrj/eslint-config-basic
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @yomukizrj/eslint-config-js@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes