+++ packages/eslint-config-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @yomukizrj/eslint-config-js
 
+## 0.3.0
+
+### Minor Changes
+
+- 修复规则冲突
+
 ## 0.2.0
 
 ### Minor Changes